requirements.txt
Django==1.10.6
gunicorn==19.7.1
mysqlclient==1.3.10
wheel==0.29.0
google-cloud-datastore==1.0.0
PyYAML==3.12
paho-mqtt==1.2.1
